Address constraints added. Bug solved.
When I relaxed the restrictions for the value of the address, it happens that the builder receives a value as a configuration that targets external registers and not memory, that is, the value is in an illegal range. That is why the checker does not start the process on the read channel because the transaction has not actually been started, while later on the write channel everything goes as it should because the tool has chosen a value in the correct range.
So the error was in the limitation that must exist because each slave has its own address space.